ref(server): Organize project services in nested modules
