+++ relay-server/src/services/processor/metrics.rs
@@ -5,8 +5,8 @@ use relay_quotas::Scoping;
 
 use crate::metrics::MetricOutcomes;
 use crate::services::outcome::Outcome;
-use crate::services::project::ProjectInfo;
-use crate::services::project_cache::BucketSource;
+use crate::services::projects::cache::BucketSource;
+use crate::services::projects::project::ProjectInfo;
 
 /// Checks if the namespace of the passed bucket is valid.
 ///
